+++ AppFairApp.xcconfig
@@ -8,16 +8,16 @@ PRODUCT_NAME = Stanza Redux
 PRODUCT_BUNDLE_IDENTIFIER = app.Stanza-Redux
 
 // The semantic version for the app matching the git tag for the release
-MARKETING_VERSION = 0.2.15
+MARKETING_VERSION = 0.3.1
 
 // The monotonically increasing number specifying the internal app version
-CURRENT_PROJECT_VERSION = 117
+CURRENT_PROJECT_VERSION = 118
 
 // The category of the app should be one of: public.app-category.business, developer-tools, education, entertainment, finance, games, graphics-design, healthcare-fitness, lifestyle, medical, music, news, photography, productivity, reference, social-networking, sports, travel, utilities, video, weather
 APP_CATEGORY = public.app-category.reference
 
 // The default tint color to apply to the generated icon, as a 6-character RGB hex string
-ICON_TINT = 187BF7
+// ICON_TINT = 187BF7
 
 // The system symbol name to embed in the generated icon instead of the default monogram
 ICON_SYMBOL = M 287.375000 114.156250 C 255.981000 114.156250 230.437469 139.700755 230.437500 171.093750 C 230.437500 202.487750 255.981028 228.031301 287.375000 228.031250 L 595.468750 228.031250 L 595.468750 210.812500 L 287.375000 210.812500 C 265.478999 210.812500 247.687632 192.990794 247.687500 171.093750 C 247.687500 149.197749 265.479034 131.374990 287.375000 131.375000 L 595.468750 131.375000 L 595.468750 114.156250 L 287.375000 114.156250 Z M 287.375000 145.250000 L 287.375000 152.250000 L 580.500000 152.250000 C 581.446000 149.799000 582.558773 147.456008 583.843750 145.250000 L 287.375000 145.250000 Z M 287.375000 168.125000 L 287.375000 175.125000 L 577.187500 175.125000 C 577.088500 173.799000 577.031213 172.449752 577.031250 171.093750 C 577.031250 170.091750 577.071045 169.110804 577.125000 168.125000 L 287.375000 168.125000 Z M 287.375000 190.968750 L 287.375000 197.968750 L 584.468750 197.968750 C 583.109752 195.767753 581.919962 193.429699 580.906250 190.968750 L 287.375000 190.968750 Z M 222.125000 237.531250 C 207.333000 237.531250 195.312378 249.582987 195.312500 264.375000 L 195.312500 324.593750 C 195.312500 339.385750 207.332967 351.406219 222.125000 351.406250 L 229.562500 351.406250 C 233.499500 351.407250 236.962622 349.345813 238.937500 346.250000 L 242.312500 346.250000 C 244.287500 349.346000 247.751214 351.406568 251.687500 351.406250 L 557.281250 351.406250 L 557.281250 334.187500 L 256.062500 334.187500 C 254.087500 331.091500 250.624803 329.031217 246.687500 329.031250 L 234.562500 329.031250 C 230.625500 329.031250 227.161371 331.092550 225.187500 334.187500 L 222.125000 334.187500 C 216.830000 334.187500 212.531217 329.888733 212.531250 324.593750 L 212.531250 264.375000 C 212.531250 259.080000 216.830017 254.749977 222.125000 254.750000 L 225.281250 254.750000 C 227.271250 257.760000 230.691357 259.749978 234.562500 259.750000 L 246.687500 259.750000 C 250.558500 259.750000 253.978860 257.759989 255.968750 254.750000 L 557.281250 254.750000 L 557.281250 237.531250 L 251.687500 237.531250 C 247.816500 237.531250 244.396140 239.521271 242.406250 242.531250 L 238.843750 242.531250 C 236.853750 239.521250 233.433510 237.531251 229.562500 237.531250 L 222.125000 237.531250 Z M 255.781250 268.156250 L 255.781250 275.156250 L 548.906250 275.156250 C 549.852250 272.705250 550.965023 270.362248 552.250000 268.156250 L 255.781250 268.156250 Z M 255.781250 291.000000 L 255.781250 298.000000 L 545.593750 298.000000 C 545.494750 296.674000 545.437504 295.324752 545.437500 293.968750 C 545.437500 292.966750 545.476237 291.985946 545.531250 291.000000 L 255.781250 291.000000 Z M 255.781250 313.875000 L 255.781250 320.875000 L 552.875000 320.875000 C 551.516000 318.674001 550.326421 316.336001 549.312500 313.875000 L 255.781250 313.875000 Z M 287.375000 362.906250 C 255.981000 362.906250 230.437469 388.450623 230.437500 419.843750 C 230.437500 451.237750 255.982015 476.781301 287.375000 476.781250 L 595.468750 476.781250 L 595.468750 459.562500 L 287.375000 459.562500 C 265.478999 459.562500 247.687500 441.740662 247.687500 419.843750 C 247.687500 397.947749 265.479024 380.125183 287.375000 380.125000 L 595.468750 380.125000 L 595.468750 362.906250 L 287.375000 362.906250 Z M 287.375000 394.000000 L 287.375000 401.000000 L 580.500000 401.000000 C 581.446000 398.549000 582.558773 396.205347 583.843750 394.000000 L 287.375000 394.000000 Z M 287.375000 416.843750 L 287.375000 423.843750 L 577.187500 423.843750 C 577.088500 422.517750 577.031213 421.199386 577.031250 419.843750 C 577.031250 418.841750 577.071045 417.829777 577.125000 416.843750 L 287.375000 416.843750 Z M 287.375000 439.718750 L 287.375000 446.718750 L 584.468750 446.718750 C 583.109752 444.517753 581.919922 442.179823 580.906250 439.718750 L 287.375000 439.718750 Z M 287.375000 439.718750
