+++ Assets.xcassets/AccentColor.colorset/Contents.json
@@ -5,7 +5,7 @@
         "color-space" : "display-p3",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "1.000",
+          "blue" : "0.361",
           "green" : "0.432",
           "red" : "0.000"
         }
