+++ Sources/App/AppContainer.swift
@@ -32,3 +32,4 @@ public extension AppContainer {
             .frame(width: 400, height: 300)
     }
 }
+
